Fix some compiler warnings

+++ src/oxygengtkcellinfo.cpp
@@ -40,7 +40,7 @@ namespace Oxygen
         points.push_back( Point( x+w, y+h-1 ) );
 
         for( std::vector<Point>::iterator iter  = points.begin(); iter != points.end() && !_path; iter++ )
-        { gtk_tree_view_get_path_at_pos( treeView, iter->x(), iter->y(), &_path, &_column, 0L, 0L ); }
+        { gtk_tree_view_get_path_at_pos( treeView, (gint)iter->x(), (gint)iter->y(), &_path, &_column, 0L, 0L ); }
 
     }
